refacto from tbz_refacto branch
